fix display of mod power

+++ src/server/scripts/Commands/cs_modify.cpp
@@ -607,7 +607,7 @@ public:
 		}
 
 		target->SetPower(value,value1);
-		handler->PSendSysMessage("Power value has been set to: ", value1);
+		handler->PSendSysMessage("Power value has been set to: ", uint32(value1));
 
 
 	}
@@ -692,7 +692,7 @@ public:
 		}
 		
 		uint32 value1 = target->GetPower(value);
-		handler->PSendSysMessage("Power value is: ", value1);
+		handler->PSendSysMessage("Power value is: ", uint32(value1));
 	}
     //Edit Player Speed
     static bool HandleModifySpeedCommand(ChatHandler* handler, const char* args)
